requirements.test.txt
pytest
pytest-cov>=4.0.0
pytest-homeassistant-custom-component
pytest-asyncio
psutil